requirements.txt
Django==1.10
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
olefile==0.44
Pillow==4.1.1
pip==9.0.1
Pygments==2.2.0
setuptools==36.0.0
six==1.10.0
wheel==0.29.0
